trying to get working fork()

+++ abis/dripos/pid_t.h
@@ -2,7 +2,7 @@
 #ifndef _ABIBITS_PID_T_H
 #define _ABIBITS_PID_T_H
 
-typedef int pid_t;
+typedef long long pid_t;
 
 #endif /* _ABIBITS_PID_T_H */
